rename: weeklyPrecipitaionUtil 파일명 및 경로 수정
