+++ frontend/src/pages/weatherBoardPage/utils/temperatureUtil.ts
